#Add breeze and email Authentication #now the users can not access to other users note and see 403 page when try to show others notes #customize 403 error page #add pagination links below of the pages

+++ notnote/app/Http/Controllers/NoteController.php
@@ -14,7 +14,10 @@ class NoteController extends Controller
     {
         // $notes = Note::query()->orderBy('created_at','desc')->get();
         // dd($notes);
-        $notes = Note::query()->orderBy('created_at','desc')->paginate();
+        $notes = Note::query()
+        ->where('user_id',request()->user()->id)
+        ->orderBy('created_at','desc')
+        ->paginate();
         return view('note.index',['notes'=>$notes]);
     }
 
@@ -38,7 +41,7 @@ class NoteController extends Controller
 
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
 
         $note = Note::create($data);
 
@@ -51,7 +54,11 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
-        return view('note.show',['note'=>$note]);
+        if($note->user_id !== request()->user()->id){
+            abort(403);
+        }else{
+            return view('note.show',['note'=>$note]);
+        }
     }
 
     /**
@@ -59,7 +66,11 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
-        return view('note.edit',['note'=>$note]);
+        if($note->user_id !== request()->user()->id){
+            abort(403);
+        }else{
+            return view('note.edit',['note'=>$note]);
+        }
     }
 
     /**
@@ -68,12 +79,16 @@ class NoteController extends Controller
     public function update(Request $request, Note $note)
     {
 
+        if($note->user_id !== request()->user()->id){
+            abort(403);
+        }
         //validate the data
         $data = $request->validate([
 
             'note' => ['required','string']
 
         ]);
+
 
         $note->update($data);
 
@@ -85,6 +100,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if($note->user_id !== request()->user()->id){
+            abort(403);
+        }
         $note->delete();
         return to_route('note.index')->with('message','Note was deleted');
